backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
sqlalchemy>=2.0.0
pydantic>=2.5.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
feedparser>=6.0.0
transformers>=4.35.0
sentence-transformers>=2.2.0
psycopg2-binary>=2.9.0
alembic>=1.12.0 